feature: autenticacao e validacao praticamente terminadas

+++ backend/src/controllers/address.controller.ts
@@ -6,11 +6,12 @@ const prisma = new PrismaClient();
 class AddressController {
 
   public async create(req: Request, res: Response) {
-    const { number, apartament, cep, userId } = req.body;
+    const { number, apartament, cep } = req.body;
+    const idUser = req.user; 
 
     try {
       const newAddress = await prisma.address.create({
-        data: { number, apartament, cep, userId }
+        data: { number, apartament, cep, user: {connect:{id:idUser}} }
       });
       return res.status(201).json(newAddress);
     } catch (error) {
@@ -31,11 +32,12 @@ class AddressController {
 
   public async update(req: Request, res: Response) {
     const { id } = req.params;
+    const idUser = req.user;
     const { number, apartament, cep } = req.body;
 
     try {
       const updatedAddress = await prisma.address.update({
-        where: { id: Number(id) },
+        where: { id: Number(id), userId: idUser },
         data: { number, apartament, cep }
       });
       return res.status(200).json(updatedAddress);
@@ -46,9 +48,10 @@ class AddressController {
 
   public async delete(req: Request, res: Response) {
     const { id } = req.params;
+    const idUser = req.user;
 
     try {
-      await prisma.address.delete({ where: { id: Number(id) } });
+      await prisma.address.delete({ where: { id: Number(id), userId: idUser } });
       return res.status(200).json({ message: "Endereço deletado com sucesso" });
     } catch (error) {
       return res.status(500).json({ messageError: "Erro deletando endereço", error});
